requirements.txt
numpy==1.21.6
opencv_python==4.6.0.66
opt_einsum==3.3.0
Pillow==9.5.0
scipy==1.7.3
torch==1.7.1
torchvision==0.8.2
transformers==4.24.0
